requirements.txt
# Dependências do Chatbot Jurídico Inteligente
# Projeto Final - Pós-Graduação iCEV - NLP

# Core ML/NLP
torch>=1.9.0
transformers>=4.20.0
sentence-transformers>=2.2.0
scikit-learn>=1.1.0
numpy>=1.21.0
pandas>=1.4.0

# Vector Database
faiss-cpu>=1.7.0

# Text Processing
langchain>=0.0.200
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1

# Web Interface
gradio>=3.35.0

# Document Processing
pdfplumber>=0.7.0
beautifulsoup4>=4.11.0
requests>=2.28.0

# Visualization and Analysis
matplotlib>=3.5.0
seaborn>=0.11.0

# Utilities
python-dateutil>=2.8.0
tqdm>=4.64.0

# Optional: For better performance
# accelerate>=0.20.0
# bitsandbytes>=0.39.0